agent/runtime/requirements.txt
bedrock-agentcore
bedrock-agentcore-starter-toolkit
strands-agents
strands-agents-tools
mcp
boto3
